requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
pyyaml>=6.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
pillow>=10.0.0
imagehash>=4.3.0
streamlit>=1.28.0
pandas>=2.0.0
